refactor: reorganize pages and improve mobile performance
