+++ src/main.ts
@@ -3,7 +3,7 @@ import { createHead } from '@vueuse/head'
 import { createApp } from 'vue'
 import App from './App.vue'
 import './assets/tailwind.css'
-import { router } from './routes'
+import { router } from './router'
 
 const app = createApp(App)
 const head = createHead()
