+++ src/router/index.ts
@@ -1,10 +1,10 @@
-import Home from '@/components/Home.vue'
+import Home from '@/views/Home.vue'
 import { createMemoryHistory, createRouter } from 'vue-router'
 
 const routes = [
     { path: '/', component: Home },
-    { path: '/contactme', component: () => import('@/components/ContactMe.vue') },
-    { path: '/portfolio', component: () => import('@/components/Portfolio.vue') },
+    { path: '/contactme', component: () => import('@/views/ContactMe.vue') },
+    { path: '/portfolio', component: () => import('@/views/Portfolio.vue') },
     { path: '/experience', component: () => import('@/components/shared/ExperienceWork.vue') }
 ]
